Subclass pkginfo's Installed class to find metadata
When a package is installed as a wheel, pkginfo cannot properly find its
metadata. To resolve this, a patch has been posted to
https://bugs.launchpad.net/pkginfo/+bug/1437570 and those fixes have been
included here.

This silences UserWarnings from pkginfo about not being able to find
PKG-INFO for a package and allows twine to accurately report pkginfo's
version once installed.

Closes #114