Expand unexpanded globs earlier than before
If we receive a glob that was unexpanded, we already expand that.
However, we were not then updating our signatures dictionary and as
such we were probably uploading packages and signatures incorrectly in
that very rare case.

This changes twine-upload to expand those globs much earlier in the
execution so we do not miss any signature data or anything else while
uploading a package.

Related to #132