Prevent flattening certain upload fields
The gpg_signature field needs to be passed to MultipartEncoder with a
tuple as the value. This was flattening it causing PyPI to receive two
parts with the same name and parsing them into a list.

The values need to otherwise be flattened because in some cases the
MultipartEncoder cannot handle a sequence value that isn't meant to be a
file-like part (e.g., platform).

Closes #137