Check response after attempting upload
We used to do this but it got lost between 1.6.x and 1.7.x. The check
to see if a package is already uploaded first is much faster but in
the event that it doesn't work, we have to attempt the upload and then
check once more if we should ignore it.

Closes gh-195