Allow setting ca_cert and client_cert in config
Right now, the settings code expects you to be able to set `ca_cert` and `client_cert` in the `pypirc` as properties of the repository (that's what `get_userpass_value` does). However, the parsing code doesn't read those properties in! This minor change fixes that issue, finally allowing the cert to be set anywhere, not just on the command line.