Add two new optional configuration locations to consider.
While earlier Python package uploading tooling only looked only for
~/.pypirc file there is no need to stick to this convention forever.
This patch introduces a secondary user level configuration path,
namespaced to this tool, and also a system-wide location that allows
system administrators to provide a sensible default.

This implementation uses the appdirs project to handle the cross
platform decision making for us (initial use case is to provide a
Windows user experience managed across a fleet of workstations equally
to Linux development environments).

Linux precedence for jsmith:

  - /home/jsmith/.config/twine/.pypirc
  - /home/jsmith/.pypirc
  - /etc/xdg/twine/.pypirc

Windows precedence for jsmith:

  - c:\Users\jsmith\AppData\Roaming\twine\.pypirc
  - c:\Users\jsmith\.pypirc
  - c:\ProgramData\twine\.pypirc

This does not attempt to teach twine to overload a lower precedence
configuration with a higher one.