Remove support for Python 2.7 and 3.4
The twine project is moving to Python 3 only.

As twine is a command-line tool, it is in a good position to move to
Python 3 only. Other command-line-only tools have begun moving to Python
3. For example: black, mypy, pylint, pypinfo, and Sphinx. The twine
command will continue to support uploading Python 2 projects to PyPI. So
this should be mostly transparent to users.

Allows for code cleanups. By removing these workarounds, the project
will be easier to maintain and reduce ongoing testing resources.

Python 2 is scheduled to be EOL on January 1, 2020.

https://www.python.org/dev/peps/pep-0373/

For an overview of projects that have dropped Python 2 support entirely,
see:

https://hugovk.github.io/drop-python/2.7/