Add explicit isort config for build package. (#966)
This is a hack to workaround an obscure edge case. As of tox>=4, running
`tox -e py` generates a `build/` directory in the project root. That
confuses isort into thinking that `build` is a first-party package,
which sorts it in the same section as `twine`. This forces `build` to be
recognized as ta third-party package.