Migrate to GitHub Issue Forms
We've had a few issues lately that haven't filled in all of the pieces
of required data. Rather than rely on humans to see and request that
information from people, let's make the tooling enforce that for us.

This also pre-labels things for us so we can better understand a new
issue.