Fix issue form syntax for GitHub validations
These seemed to work when we created them but GitHub has become more
srict in its validation. That's a whole lot of fun when there's nothing
to verify these locally. Naturally, the top of their "common validation
errors" page says "Issue forms are currently in beta and subject to
change" so that probably explains it just fine.