Remove tests/alt-fixtures/twine-1.5.0-py2.py3-none-any.whl
This test file was added in 8b8790b93ce7 (Add tests to ensure version
parsing does not regress, part of #147) with the intent of testing the
target Python version from wheel file names.

However, the filename is identical to the one of the other text
fixture tests/fixtures/twine-1.5.0-py2.py3-none-any.whl thus it does
not appear to add any additional coverage to the code extracting the
target Python version.