Simplify generation of test packages used in test_check
These tests used `build` to invoke the `setuptools.build_meta` build
backend on a package directory generated on the fly. The tests are
only interested in the content of the package metadata and use
variations of `setup.cfg` to generate the desired metadata.

This can be simplified to the direct generation of sdist archives with
synthetic `PKG-INFO` metadata files. This makes the actual metadata
the tests are checking obvious and avoid two test dependencies.

The tests simplification highlighted that two tests are actually
testing the exact same thing. Remove one.