bugfix: utils: catch configparser.Error
This catches any configparser exceptions when parsing the user's
pypirc file. Without this the user would see an uncontrolled
exception trace, which in turn could sometimes leak sensitive
configuration fields.

The "fix" here is to mask the configparser exception entirely
with twine's own `InvalidConfiguration` exception. This avoids
the leakage risk, but it also means that the errors we return to
users are slightly less specific in terms of where their invalid
syntax is. We offset that by including a link to the pypirc docs.

Fixes #1233.

Signed-off-by: William Woodruff <william@yossarian.net>