Automatically replace Trusted Publishing Tokens
Tokens obtained via the Trusted Publishing flow are only valid for a
short period of time. As such, projects using Trusted Publishing and
trying to upload a large number of artifacts (or potentially a large number
of large artifacts) could run out of time before the token expires.

This is a first pass at solving the problem by replacing the token when
it's within 5 minutes of expiry. This also adds optional parsing from
the Warehouse API which does not currently return the expiration
timestamp. If that's unavailable, we guesstimate the remaining validity.

Closes gh-1246