feat(package): remove MD5 hashing entirely
This removes MD5 from HashManager, which in turn means
we no longer supply a `md5_digest` in the metadata when uploading
to a repository. This is consistent with PyPI's own expectation
that MD5 is optional.

In terms of security this has no impact as MD5 is considered
insecure for integrity purposes and PyPI already uniformly
uses SHA-256 and BLAKE2 for package integrity.