Add --ignore-http-status option
This is effectively a replacement for --skip-existing with non-PyPI
repositories.  It is not allowed when using [Test]PyPI, and is instead a
way for users of other repositories to choose which status(es) can be
ignored.

This commit also rewords the UnsupportedConfiguration message a bit;
instead of asserting that the configured repository does not support a
particular feature (which is not actually checked), instead assert that
Twine does not support using the feature(s) with the configured
repository.